# dcrpg: combine spent and unspent totals queries into one