--- a/db/dcrpg/internal/addrstmts.go
+++ b/db/dcrpg/internal/addrstmts.go
@@ -12,14 +12,13 @@ const (
 	InsertAddressRowReturnID = `WITH inserting AS (` +
 		insertAddressRow0 +
 		`ON CONFLICT (tx_vin_vout_row_id, address, is_funding) DO UPDATE
-		SET address = NULL WHERE FALSE
-		RETURNING id
+			SET address = NULL WHERE FALSE
+			RETURNING id
 		)
 		SELECT id FROM inserting
 		UNION  ALL
 		SELECT id FROM addresses
-		WHERE  address = $1, is_funding = TRUE
-		AND tx_vin_vout_row_id = $5
+		WHERE  address = $1 AND is_funding = TRUE AND tx_vin_vout_row_id = $5
 		LIMIT  1;`
 
 	// SelectSpendingTxsByPrevTx = `SELECT id, tx_hash, tx_index, prev_tx_index FROM vins WHERE prev_tx_hash=$1;`
@@ -48,19 +47,19 @@ const (
 	SelectAddressRecvCount    = `SELECT COUNT(*) FROM addresses WHERE address=$1 AND valid_mainchain = TRUE;`
 
 	SelectAddressesAllTxn = `SELECT
-		transactions.tx_hash,
-		block_height
-	FROM
-		addresses
-		INNER JOIN
-			transactions
-			ON addresses.tx_hash = transactions.tx_hash
-			AND is_mainchain = TRUE AND is_valid=TRUE
-	WHERE
-		address = ANY($1) AND valid_mainchain=true
-	ORDER BY
-		time DESC,
-		transactions.tx_hash ASC;`
+			transactions.tx_hash,
+			block_height
+		FROM
+			addresses
+			INNER JOIN
+				transactions
+				ON addresses.tx_hash = transactions.tx_hash
+				AND is_mainchain = TRUE AND is_valid=TRUE
+		WHERE
+			address = ANY($1) AND valid_mainchain=true
+		ORDER BY
+			time DESC,
+			transactions.tx_hash ASC;`
 
 	SelectAddressUnspentCountANDValue = `SELECT COUNT(*), SUM(value) FROM addresses
 	    WHERE address = $1 AND is_funding = TRUE AND matching_tx_hash = '' AND valid_mainchain = TRUE;`
@@ -68,22 +67,53 @@ const (
 	SelectAddressSpentCountANDValue = `SELECT COUNT(*), SUM(value) FROM addresses
 		WHERE address = $1 AND is_funding = FALSE AND matching_tx_hash != '' AND valid_mainchain = TRUE;`
 
-	SelectAddressesMergedSpentCount = `SELECT COUNT( distinct tx_hash ) FROM addresses
+	SelectAddressesMergedSpentCount = `SELECT COUNT( DISTINCT tx_hash ) FROM addresses
 		WHERE address = $1 AND is_funding = FALSE AND valid_mainchain = TRUE;`
 
+	// SelectAddressSpentUnspentCountAndValue gets the number and combined spent
+	// and unspent outpoints for the given address. The key is the "GROUP BY
+	// is_funding, matching_tx_hash=''" part of the statement that gets the data
+	// for the combinations of is_funding (boolean) and matching_tx_hash=''
+	// (boolean). There should never be any with is_funding=true where
+	// matching_tx_hash is empty, thus there are three rows in the output. For
+	// example, the first row is the spending transactions that must have
+	// matching_tx_hash set, the second row the the funding transactions for the
+	// first row (notice the equal count and sum), and the third row are the
+	// unspent outpoints that are is_funding=true but with an empty
+	// matching_tx_hash:
+	//
+	// count  |      sum       | is_funding | all_empty_matching | no_empty_matching
+	// --------+----------------+------------+--------------------+--------------------
+	//   45150 | 12352318108368 | f          | f                  | t
+	//   45150 | 12352318108368 | t          | f                  | t
+	//  229145 | 55875634749104 | t          | t                  | f
+	// (3 rows)
+	//
+	// Since part of the grouping is on "matching_tx_hash = ''", what is
+	// logically "any" empty matching is actually no_empty_matching.
+	SelectAddressSpentUnspentCountAndValue = `SELECT COUNT(*),
+			SUM(value),
+			is_funding,
+			BOOL_AND(matching_tx_hash = '') AS all_empty_matching
+			-- NOT BOOL_AND(matching_tx_hash = '') AS no_empty_matching
+		FROM addresses
+		WHERE address = $1 AND valid_mainchain = TRUE
+		GROUP BY is_funding, matching_tx_hash=''  -- separate spent and unspent
+		ORDER BY count, is_funding;`
+
 	SelectAddressUnspentWithTxn = `SELECT
-		addresses.address,
-		addresses.tx_hash,
-		addresses.value,
-		transactions.block_height,
-		addresses.block_time,
-		tx_vin_vout_index,
-		pkscript
+			addresses.address,
+			addresses.tx_hash,
+			addresses.value,
+			transactions.block_height,
+			addresses.block_time,
+			tx_vin_vout_index,
+			pkscript
 		FROM addresses
 		JOIN transactions ON
 			addresses.tx_hash = transactions.tx_hash
-		JOIN vouts on addresses.tx_hash = vouts.tx_hash AND addresses.tx_vin_vout_index=vouts.tx_index
-		WHERE addresses.address=$1 AND addresses.is_funding = TRUE and addresses.matching_tx_hash = '' AND valid_mainchain = TRUE
+		JOIN vouts ON addresses.tx_hash = vouts.tx_hash AND addresses.tx_vin_vout_index=vouts.tx_index
+		WHERE addresses.address=$1 AND addresses.is_funding = TRUE AND addresses.matching_tx_hash = '' AND valid_mainchain = TRUE
 		ORDER BY addresses.block_time DESC;`
 
 	SelectAddressLimitNByAddress = `SELECT ` + addrsColumnNames + ` FROM addresses
@@ -93,9 +123,11 @@ const (
 		` FROM addresses WHERE address=$1 AND valid_mainchain = TRUE)
 		SELECT * FROM these ORDER BY block_time DESC LIMIT $2 OFFSET $3;`
 
-	SelectAddressMergedDebitView = `SELECT tx_hash, valid_mainchain, block_time, sum(value),
-		COUNT(*) FROM addresses WHERE address=$1 AND is_funding = FALSE
-		GROUP BY (tx_hash, valid_mainchain, block_time) ORDER BY block_time DESC LIMIT $2 OFFSET $3;`
+	SelectAddressMergedDebitView = `SELECT tx_hash, valid_mainchain, block_time, sum(value), COUNT(*)
+		FROM addresses
+		WHERE address=$1 AND is_funding = FALSE          -- spending transactions
+		GROUP BY (tx_hash, valid_mainchain, block_time)  -- merging common transactions in same valid mainchain block
+		ORDER BY block_time DESC LIMIT $2 OFFSET $3;`
 
 	SelectAddressDebitsLimitNByAddress = `SELECT ` + addrsColumnNames + `
 		FROM addresses WHERE address=$1 AND is_funding = FALSE AND valid_mainchain = TRUE