Add `SelectAddressSpentUnspentCountAndValue` query that gets the count and total value for spent and unspent address table rows for a given address.  It uses `GROUP BY (is_funding, matching_tx_hash='')` with aggregate functions to do it efficiently in a single scan on the addresses table.
Scanning the result, `isFunding && noMatchingTx` means unspent, `!isFunding && !noMatchingTx` means spent, while `!isFunding && noMatchingTx` is not expected so an error is logged.

Timings on one machine (project fund):

| query  | timing   |
|---|---|
| new unified query  | 230 ms  |
| old unspent query  | 380 ms  |
| old spent query | 227 ms |
| old spent+unspent queries | 607 ms |

That's a 62% reduction in time (down to 38% of the original query cost) to get both spent and unspent totals for the project fund.

Also clean-up some query formatting.
Fix the conditional on the (unused) `InsertAddressRowReturnID` query.